+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/head/structure/Head.java
@@ -1,6 +1,8 @@
 package io.github.sakurawald.fuji.module.initializer.head.structure;
 
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.text.Text;
@@ -9,16 +11,18 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.UUID;
 
+@Data
+@NoArgsConstructor
 public class Head {
 
-    public final @NotNull String name;
+    public @NotNull String name;
 
-    public final UUID uuid;
+    public UUID uuid;
 
-    public final String value;
+    public String value;
 
     @Nullable
-    public final String tags;
+    public String tags;
 
     public Head(@NotNull String name, UUID uuid, String value, @Nullable String tags) {
         this.name = name;